pyproject.toml
[tool.poetry]
name = "project-chiral-qa-service"
version = "0.1.0"
description = "问答系统微服务"
authors = ["王博伟 <wozluohd@gmail.com>"]
readme = "README.md"
packages = [{ include = "project_chiral_qa_service" }]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.25.1"
torch = "^1.13.1"
pika = "^1.3.1"
faiss-gpu = "^1.7.2"
farm-haystack = {extras = ["faiss-gpu"], version = "^1.13.2"}


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
